+++ solidity/contracts/converter/types/liquidity-pool-v1/LiquidityPoolV1Converter.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: SEE LICENSE IN LICENSE
 pragma solidity 0.6.12;
 import "../../LiquidityPoolConverter.sol";
-import "../../../token/interfaces/ISmartToken.sol";
+import "../../../token/interfaces/IDSToken.sol";
 import "../../../utility/Types.sol";
 
 /**
@@ -32,7 +32,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
       * deprecated, use `TokenRateUpdate` from version 28 and up
       *
       * @param  _connectorToken     reserve token
-      * @param  _tokenSupply        smart token supply
+      * @param  _tokenSupply        pool token supply
       * @param  _connectorBalance   reserve balance
       * @param  _connectorWeight    reserve weight
     */
@@ -51,7 +51,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
       * @param  _maxConversionFee   maximum conversion fee, represented in ppm
     */
     constructor(
-        ISmartToken _token,
+        IDSToken _token,
         IContractRegistry _registry,
         uint32 _maxConversionFee
     )
@@ -281,7 +281,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         }
 
         // get the total supply
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
 
         // transfer from the user an equally-worth amount of each one of the reserve tokens
         uint256 amount = addLiquidityToPool(_reserveTokens, _reserveAmounts, totalSupply);
@@ -290,7 +290,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         require(amount >= _minReturn, "ERR_RETURN_TOO_LOW");
 
         // issue the tokens to the user
-        ISmartToken(address(anchor)).issue(msg.sender, amount);
+        IDSToken(address(anchor)).issue(msg.sender, amount);
 
         // return the amount of pool tokens issued
         return amount;
@@ -316,10 +316,10 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         verifyLiquidityInput(_reserveTokens, _reserveMinReturnAmounts, _amount);
 
         // get the total supply BEFORE destroying the user tokens
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
 
         // destroy the user tokens
-        ISmartToken(address(anchor)).destroy(msg.sender, _amount);
+        IDSToken(address(anchor)).destroy(msg.sender, _amount);
 
         // transfer to the user an equivalent amount of each one of the reserve tokens
         return removeLiquidityFromPool(_reserveTokens, _reserveMinReturnAmounts, totalSupply, _amount);
@@ -344,7 +344,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         syncReserveBalances();
         reserves[ETH_RESERVE_ADDRESS].balance = reserves[ETH_RESERVE_ADDRESS].balance.sub(msg.value);
 
-        uint256 supply = ISmartToken(address(anchor)).totalSupply();
+        uint256 supply = IDSToken(address(anchor)).totalSupply();
         IBancorFormula formula = IBancorFormula(addressOf(BANCOR_FORMULA));
 
         // iterate through the reserve tokens and transfer a percentage equal to the weight between
@@ -384,7 +384,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         }
 
         // issue new funds to the caller in the pool token
-        ISmartToken(address(anchor)).issue(msg.sender, _amount);
+        IDSToken(address(anchor)).issue(msg.sender, _amount);
 
         // return the amount of pool tokens issued
         return _amount;
@@ -407,8 +407,8 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
     {
         require(_amount > 0, "ERR_ZERO_AMOUNT");
 
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
-        ISmartToken(address(anchor)).destroy(msg.sender, _amount);
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
+        IDSToken(address(anchor)).destroy(msg.sender, _amount);
 
         uint256[] memory reserveMinReturnAmounts = new uint256[](reserveTokens.length);
         for (uint256 i = 0; i < reserveMinReturnAmounts.length; i++)
@@ -436,7 +436,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
     {
         uint256[] memory reserveAmounts = new uint256[](_reserveTokens.length);
 
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
         IBancorFormula formula = IBancorFormula(addressOf(BANCOR_FORMULA));
         uint256 amount = formula.fundSupplyAmount(totalSupply, reserves[_reserveTokens[_reserveTokenIndex]].balance, reserveRatio, _reserveAmount);
 
@@ -462,7 +462,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         view
         returns (uint256)
     {
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
         IBancorFormula formula = IBancorFormula(addressOf(BANCOR_FORMULA));
         return formula.fundSupplyAmount(totalSupply, reserves[_reserveToken].balance, reserveRatio, _reserveAmount);
     }
@@ -480,7 +480,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
         view
         returns (uint256[] memory)
     {
-        uint256 totalSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 totalSupply = IDSToken(address(anchor)).totalSupply();
         IBancorFormula formula = IBancorFormula(addressOf(BANCOR_FORMULA));
         return removeLiquidityReserveAmounts(_amount, _reserveTokens, totalSupply, formula);
     }
@@ -742,7 +742,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
       * @param _targetToken address of the target reserve token
     */
     function dispatchTokenRateUpdateEvents(IERC20Token _sourceToken, IERC20Token _targetToken) private {
-        uint256 poolTokenSupply = ISmartToken(address(anchor)).totalSupply();
+        uint256 poolTokenSupply = IDSToken(address(anchor)).totalSupply();
         uint256 sourceReserveBalance = reserveBalance(_sourceToken);
         uint256 targetReserveBalance = reserveBalance(_targetToken);
         uint32 sourceReserveWeight = reserves[_sourceToken].weight;
@@ -771,7 +771,7 @@ contract LiquidityPoolV1Converter is LiquidityPoolConverter {
       * @param _reserveWeight   reserve weight
     */
     function dispatchPoolTokenRateUpdateEvent(uint256 _poolTokenSupply, IERC20Token _reserveToken, uint256 _reserveBalance, uint32 _reserveWeight) private {
-        emit TokenRateUpdate(ISmartToken(address(anchor)), _reserveToken, _reserveBalance.mul(PPM_RESOLUTION), _poolTokenSupply.mul(_reserveWeight));
+        emit TokenRateUpdate(IDSToken(address(anchor)), _reserveToken, _reserveBalance.mul(PPM_RESOLUTION), _poolTokenSupply.mul(_reserveWeight));
     }
 
     /**